+++ lib/src/controls/inputs/toggle_button.dart
@@ -55,7 +55,8 @@ class ToggleButton extends StatelessWidget {
     properties.add(DiagnosticsProperty<ToggleButtonThemeData>('style', style));
     properties.add(StringProperty('semanticLabel', semanticLabel));
     properties.add(ObjectFlagProperty<FocusNode>.has('focusNode', focusNode));
-    properties.add(FlagProperty('autofocus', value: autofocus));
+    properties.add(
+        FlagProperty('autofocus', value: autofocus, ifFalse: 'manual focus'));
   }
 
   @override