+++ lib/src/controls/surfaces/calendar/calendar_view.dart
@@ -1,1 +1,24 @@
+import 'package:fluent_ui/fluent_ui.dart';
 
+// TODO: Implement calendar view (https://github.com/bdlukaa/fluent_ui/issues/18)
+
+/// A calendar view lets a user view and interact with a
+/// calendar that they can navigate by month, year, or
+/// decade. A user can select a single date or a range of
+/// dates. It doesn't have a picker surface and the calendar
+/// is always visible.
+///
+/// ![CalendarView Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/calendar-view-3-views.png)
+///
+/// See also:
+///   - [DatePicker]
+///   - [TimePicker]
+///   - [CalendarDatePicker]
+class CalendarView extends StatelessWidget {
+  const CalendarView({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox.shrink();
+  }
+}