+++ lib/src/styles/color.dart
@@ -1,5 +1,4 @@
 import 'package:fluent_ui/fluent_ui.dart';
-import 'package:flutter/cupertino.dart' as c;
 
 /// All the fluent colors
 class Colors {