+++ lib/src/styles/motion/page_transitions.dart
@@ -1,7 +1,14 @@
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 
+/// Entrance is a combination of a slide up animation and a fade in
+/// animation for the incoming content. Use page refresh when the
+/// user is taken to the top of a navigational stack, such as navigating
+/// between tabs or left-nav items.
+///
+/// This animation is used by default by [NavigationPanelBody]
 class EntrancePageTransition extends StatelessWidget {
+  /// Creates an entrance page transition
   const EntrancePageTransition({
     Key? key,
     required this.child,
@@ -10,9 +17,16 @@ class EntrancePageTransition extends StatelessWidget {
     this.reverse = false,
   }) : super(key: key);
 
+  /// The widget to be animated
   final Widget child;
+
+  /// The animation to drive this transition
   final Animation<double> animation;
+
+  /// Whether the animation should be done vertically or horizontally
   final bool vertical;
+
+  /// Whether the animation should be done from the left or from the right
   final bool reverse;
 
   @override
@@ -27,6 +41,7 @@ class EntrancePageTransition extends StatelessWidget {
     properties.add(FlagProperty(
       'reverse',
       value: reverse,
+      ifFalse: 'from right',
       defaultValue: false,
     ));
     properties.add(PercentProperty(
@@ -53,14 +68,20 @@ class EntrancePageTransition extends StatelessWidget {
   }
 }
 
+/// Use drill when users navigate deeper into an app, such as
+/// displaying more information after selecting an item.
 class DrillInPageTransition extends StatelessWidget {
+  /// Creates a drill in page transition.
   const DrillInPageTransition({
     Key? key,
     required this.child,
     required this.animation,
   }) : super(key: key);
 
+  /// The widget to be animated
   final Widget child;
+
+  /// The animation to drive this transition
   final Animation<double> animation;
 
   @override
@@ -85,7 +106,13 @@ class DrillInPageTransition extends StatelessWidget {
   }
 }
 
+/// Use horizontal slide to show that sibling pages appear
+/// next to each other. [NavigationPanel] automatically uses
+/// this animation for top nav, but if you are building your
+/// own horizontal navigation experience, then you can implement
+/// horizontal slide with SlideNavigationTransitionInfo.
 class HorizontalSlidePageTransition extends StatelessWidget {
+  /// Creates a horizontal slide page transition.
   const HorizontalSlidePageTransition({
     Key? key,
     required this.child,
@@ -93,8 +120,13 @@ class HorizontalSlidePageTransition extends StatelessWidget {
     this.fromLeft = true,
   }) : super(key: key);
 
+  /// The widget to be animated
   final Widget child;
+
+  /// The animation to drive this transition
   final Animation<double> animation;
+
+  /// Whether this animation should be done from the left or not
   final bool fromLeft;
 
   @override
@@ -109,7 +141,7 @@ class HorizontalSlidePageTransition extends StatelessWidget {
       'fromLeft',
       value: fromLeft,
       defaultValue: true,
-      ifFalse: 'fromRight',
+      ifFalse: 'from right',
     ));
   }
 
@@ -135,10 +167,13 @@ class HorizontalSlidePageTransition extends StatelessWidget {
   }
 }
 
+/// To avoid playing any animation during navigation, use this
+/// animation.
 class SuppressPageTransition extends StatelessWidget {
   const SuppressPageTransition({Key? key, required this.child})
       : super(key: key);
 
+  /// The widget to be animation
   final Widget child;
 
   @override